move the OpenAiChatOptions.java out of the api package
