+++ spring-ai-spring-boot-autoconfigure/src/main/java/org/springframework/ai/autoconfigure/openai/OpenAiChatProperties.java
@@ -16,7 +16,7 @@
 
 package org.springframework.ai.autoconfigure.openai;
 
-import org.springframework.ai.openai.api.OpenAiChatOptions;
+import org.springframework.ai.openai.OpenAiChatOptions;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.NestedConfigurationProperty;
